refactor(#249): improve error handling type safety
