+++ apps/frontend/src/features/auth/components/SignUpForm.tsx
@@ -93,7 +93,7 @@ export const SignUpForm = () => {
       resetForm();
       navigate('/');
     } catch {
-      // publicClient interceptor
+      // 에러 토스트는 인터셉터가 처리
     } finally {
       setIsSubmitting(false);
     }
